refactor: rename P2 violations in util, tabular-source, gis-sdk, plugin-base, plugin-ui-host (#1203)
17 files renamed across 5 packages:
- util: db-name→dbNameUtils, format→formatUtils, validation→validationUtils, webCrypto→webCryptoUtils
- tabular-source: capability→capabilityUtils, ports→portTypes, registry→registryUtils, store→storeTypes
- gis-sdk: config→configTypes, geometryEngine→geometryEngineUtils, tileStore→tileStoreTypes
- plugin-base: plugin-definition→pluginDefinitionTypes, plugin-manifest→pluginManifestTypes, plugin-metadata→pluginMetadataTypes, registry→registryTypes
- plugin-ui-host: PluginDialogShell.types→pluginDialogShellTypes, frame-state→frameStateUtils

Rollback: git revert <hash> --no-edit
