+++ packages/gis-sdk/src/configTypes.ts
