+++ packages/gis-sdk/src/tileStoreTypes.ts
