+++ packages/plugin-base/src/index.ts
@@ -13,7 +13,7 @@ export * from './types/BaseSearchCriteria.js';
 export * from './types/EntityLifecycleHooks.js';
 // export * from './types/PluginExtensionAPI.js';
 export * from './types/PluginLifecycleAPI.js';
-export * from './types/plugin-definition.js';
-export * from './types/plugin-manifest.js';
-export * from './types/plugin-metadata.js';
-export * from './types/registry.js';
+export * from './types/pluginDefinitionTypes.js';
+export * from './types/pluginManifestTypes.js';
+export * from './types/pluginMetadataTypes.js';
+export * from './types/registryTypes.js';
