+++ packages/plugin-base/src/types/PackageJson.ts
@@ -1,4 +1,4 @@
-import type { PluginDefinition } from './plugin-definition.js';
+import type { PluginDefinition } from './pluginDefinitionTypes.js';
 
 /**
  * Package.json structure for plugin discovery
