+++ packages/plugin-base/src/types/pluginDefinitionTypes.ts
