+++ packages/plugin-base/src/types/pluginMetadataTypes.ts
