+++ packages/plugin-base/src/types/registryTypes.ts
