+++ packages/plugin-ui-host/src/headless/pluginDialogShellTypes.ts
