+++ packages/plugin-ui-host/src/headless/usePluginDialogController/frameStateUtils.ts
