+++ packages/plugin-ui-host/src/index.ts
@@ -1,7 +1,7 @@
 export type { PluginDialogFooterPrimaryButtonOptions } from './headless/components/PluginDialogFooter.js';
 export { PluginDialogFooter } from './headless/components/PluginDialogFooter.js';
 export { PluginDialogHeader } from './headless/components/PluginDialogHeader.js';
-export type { PluginDialogShellProps } from './headless/PluginDialogShell.types.js';
+export type { PluginDialogShellProps } from './headless/pluginDialogShellTypes.js';
 export { PluginDialogShell } from './headless/PluginDialogShell.js';
 export type { PluginDialogFooterOptions } from './headless/usePluginDialogController.js';
 export type { PluginDialogHostProps } from './PluginDialogHost.js';
