+++ packages/tabular-source/src/capabilityUtils.ts
