+++ packages/tabular-source/src/index.ts
@@ -1,9 +1,9 @@
 export * from './types.js';
-export * from './ports.js';
-export * from './registry.js';
+export * from './portTypes.js';
+export * from './registryUtils.js';
 export * from './TabularService.js';
-export * from './capability.js';
-export * from './store.js';
+export * from './capabilityUtils.js';
+export * from './storeTypes.js';
 export * from './processor.js';
 export * from './processors/createColumnRenameProcessor.js';
 export * from './processors/createNumberCoerceProcessor.js';
